Add swagger annotation delegates
