+++ springdoc-openapi-common/src/main/java/org/springdoc/core/customizers/DataRestDelegatingMethodParameterCustomizer.java
@@ -31,22 +31,17 @@ import java.util.Optional;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import io.swagger.v3.core.util.ObjectMapperFactory;
-import io.swagger.v3.oas.annotations.ExternalDocumentation;
 import io.swagger.v3.oas.annotations.Parameter;
-import io.swagger.v3.oas.annotations.enums.Explode;
-import io.swagger.v3.oas.annotations.enums.ParameterIn;
-import io.swagger.v3.oas.annotations.enums.ParameterStyle;
-import io.swagger.v3.oas.annotations.extensions.Extension;
 import io.swagger.v3.oas.annotations.media.ArraySchema;
-import io.swagger.v3.oas.annotations.media.Content;
-import io.swagger.v3.oas.annotations.media.DiscriminatorMapping;
-import io.swagger.v3.oas.annotations.media.ExampleObject;
 import io.swagger.v3.oas.annotations.media.Schema;
 import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.reflect.FieldUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springdoc.core.DelegatingMethodParameter;
+import org.springdoc.core.delegates.ArraySchemaDelegate;
+import org.springdoc.core.delegates.ParameterDelegate;
+import org.springdoc.core.delegates.SchemaDelegate;
 import org.springdoc.core.providers.RepositoryRestConfigurationProvider;
 import org.springdoc.core.providers.SpringDataWebPropertiesProvider;
 
@@ -123,180 +118,15 @@ public class DataRestDelegatingMethodParameterCustomizer implements DelegatingMe
 		try {
 			field = methodParameter.getContainingClass().getDeclaredField(parameterName);
 			Parameter parameter = field.getAnnotation(Parameter.class);
-			parameterNew = new Parameter() {
-				@Override
-				public Class<? extends Annotation> annotationType() {
-					return parameter.annotationType();
-				}
-
+			parameterNew = new ParameterDelegate(parameter) {
 				@Override
 				public String name() {
 					return getName(parameterName, parameter.name());
 				}
 
 				@Override
-				public ParameterIn in() {
-					return parameter.in();
-				}
-
-				@Override
-				public String description() {
-					return parameter.description();
-				}
-
-				@Override
-				public boolean required() {
-					return parameter.required();
-				}
-
-				@Override
-				public boolean deprecated() {
-					return parameter.deprecated();
-				}
-
-				@Override
-				public boolean allowEmptyValue() {
-					return parameter.allowEmptyValue();
-				}
-
-				@Override
-				public ParameterStyle style() {
-					return parameter.style();
-				}
-
-				@Override
-				public Explode explode() {
-					return parameter.explode();
-				}
-
-				@Override
-				public boolean allowReserved() {
-					return parameter.allowReserved();
-				}
-
-				@Override
 				public Schema schema() {
-					return new Schema() {
-
-						@Override
-						public Class<? extends Annotation> annotationType() {
-							return parameter.schema().annotationType();
-						}
-
-						@Override
-						public Class<?> implementation() {
-							return parameter.schema().implementation();
-						}
-
-						@Override
-						public Class<?> not() {
-							return parameter.schema().not();
-						}
-
-						@Override
-						public Class<?>[] oneOf() {
-							return parameter.schema().oneOf();
-						}
-
-						@Override
-						public Class<?>[] anyOf() {
-							return parameter.schema().anyOf();
-						}
-
-						@Override
-						public Class<?>[] allOf() {
-							return parameter.schema().allOf();
-						}
-
-						@Override
-						public String name() {
-							return parameter.schema().name();
-						}
-
-						@Override
-						public String title() {
-							return parameter.schema().title();
-						}
-
-						@Override
-						public double multipleOf() {
-							return parameter.schema().multipleOf();
-						}
-
-						@Override
-						public String maximum() {
-							return parameter.schema().maximum();
-						}
-
-						@Override
-						public boolean exclusiveMaximum() {
-							return parameter.schema().exclusiveMaximum();
-						}
-
-						@Override
-						public String minimum() {
-							return parameter.schema().minimum();
-						}
-
-						@Override
-						public boolean exclusiveMinimum() {
-							return parameter.schema().exclusiveMaximum();
-						}
-
-						@Override
-						public int maxLength() {
-							return parameter.schema().maxLength();
-						}
-
-						@Override
-						public int minLength() {
-							return parameter.schema().minLength();
-						}
-
-						@Override
-						public String pattern() {
-							return parameter.schema().pattern();
-						}
-
-						@Override
-						public int maxProperties() {
-							return parameter.schema().maxProperties();
-						}
-
-						@Override
-						public int minProperties() {
-							return parameter.schema().minProperties();
-						}
-
-						@Override
-						public String[] requiredProperties() {
-							return parameter.schema().requiredProperties();
-						}
-
-						@Override
-						public boolean required() {
-							return parameter.schema().required();
-						}
-
-						@Override
-						public String description() {
-							return parameter.schema().description();
-						}
-
-						@Override
-						public String format() {
-							return parameter.schema().format();
-						}
-
-						@Override
-						public String ref() {
-							return parameter.schema().ref();
-						}
-
-						@Override
-						public boolean nullable() {
-							return parameter.schema().nullable();
-						}
+					return new SchemaDelegate(parameter.schema()) {
 
 						@Override
 						public boolean readOnly() {
@@ -309,73 +139,8 @@ public class DataRestDelegatingMethodParameterCustomizer implements DelegatingMe
 						}
 
 						@Override
-						public AccessMode accessMode() {
-							return parameter.schema().accessMode();
-						}
-
-						@Override
-						public String example() {
-							return parameter.schema().example();
-						}
-
-						@Override
-						public ExternalDocumentation externalDocs() {
-							return parameter.schema().externalDocs();
-						}
-
-						@Override
-						public boolean deprecated() {
-							return parameter.schema().deprecated();
-						}
-
-						@Override
-						public String type() {
-							return parameter.schema().type();
-						}
-
-						@Override
-						public String[] allowableValues() {
-							return parameter.schema().allowableValues();
-						}
-
-						@Override
 						public String defaultValue() {
 							return getDefaultValue(parameterName, pageableDefault, parameter.schema().defaultValue());
-						}
-
-						@Override
-						public String discriminatorProperty() {
-							return parameter.schema().discriminatorProperty();
-						}
-
-						@Override
-						public DiscriminatorMapping[] discriminatorMapping() {
-							return parameter.schema().discriminatorMapping();
-						}
-
-						@Override
-						public boolean hidden() {
-							return parameter.schema().hidden();
-						}
-
-						@Override
-						public boolean enumAsRef() {
-							return parameter.schema().enumAsRef();
-						}
-
-						@Override
-						public Class<?>[] subTypes() {
-							return parameter.schema().subTypes();
-						}
-
-						@Override
-						public Extension[] extensions() {
-							return parameter.schema().extensions();
-						}
-
-						@Override
-						public AdditionalPropertiesValue additionalProperties() {
-							return parameter.schema().additionalProperties();
 						}
 					};
 				}
@@ -383,141 +148,11 @@ public class DataRestDelegatingMethodParameterCustomizer implements DelegatingMe
 				@Override
 				public ArraySchema array() {
 					ArraySchema arraySchema = parameter.array();
-					return new ArraySchema() {
-						@Override
-						public Class<? extends Annotation> annotationType() {
-							return arraySchema.annotationType();
-						}
-
-						@Override
-						public Schema schema() {
-							return arraySchema.schema();
-						}
-
+					return new ArraySchemaDelegate(arraySchema) {
 						@Override
 						public Schema arraySchema() {
 							Schema schema = arraySchema.arraySchema();
-							return new Schema() {
-
-								@Override
-								public Class<? extends Annotation> annotationType() {
-									return schema.annotationType();
-								}
-
-								@Override
-								public Class<?> implementation() {
-									return schema.implementation();
-								}
-
-								@Override
-								public Class<?> not() {
-									return schema.not();
-								}
-
-								@Override
-								public Class<?>[] oneOf() {
-									return schema.oneOf();
-								}
-
-								@Override
-								public Class<?>[] anyOf() {
-									return schema.anyOf();
-								}
-
-								@Override
-								public Class<?>[] allOf() {
-									return schema.allOf();
-								}
-
-								@Override
-								public String name() {
-									return schema.name();
-								}
-
-								@Override
-								public String title() {
-									return schema.title();
-								}
-
-								@Override
-								public double multipleOf() {
-									return schema.multipleOf();
-								}
-
-								@Override
-								public String maximum() {
-									return schema.maximum();
-								}
-
-								@Override
-								public boolean exclusiveMaximum() {
-									return schema.exclusiveMaximum();
-								}
-
-								@Override
-								public String minimum() {
-									return schema.minimum();
-								}
-
-								@Override
-								public boolean exclusiveMinimum() {
-									return schema.exclusiveMinimum();
-								}
-
-								@Override
-								public int maxLength() {
-									return schema.maxLength();
-								}
-
-								@Override
-								public int minLength() {
-									return schema.minLength();
-								}
-
-								@Override
-								public String pattern() {
-									return schema.pattern();
-								}
-
-								@Override
-								public int maxProperties() {
-									return schema.maxProperties();
-								}
-
-								@Override
-								public int minProperties() {
-									return schema.minProperties();
-								}
-
-								@Override
-								public String[] requiredProperties() {
-									return schema.requiredProperties();
-								}
-
-								@Override
-								public boolean required() {
-									return schema.required();
-								}
-
-								@Override
-								public String description() {
-									return schema.description();
-								}
-
-								@Override
-								public String format() {
-									return schema.format();
-								}
-
-								@Override
-								public String ref() {
-									return schema.ref();
-								}
-
-								@Override
-								public boolean nullable() {
-									return schema.nullable();
-								}
+							return new SchemaDelegate(schema) {
 
 								@Override
 								public boolean readOnly() {
@@ -530,127 +165,12 @@ public class DataRestDelegatingMethodParameterCustomizer implements DelegatingMe
 								}
 
 								@Override
-								public AccessMode accessMode() {
-									return schema.accessMode();
-								}
-
-								@Override
-								public String example() {
-									return schema.example();
-								}
-
-								@Override
-								public ExternalDocumentation externalDocs() {
-									return schema.externalDocs();
-								}
-
-								@Override
-								public boolean deprecated() {
-									return schema.deprecated();
-								}
-
-								@Override
-								public String type() {
-									return schema.type();
-								}
-
-								@Override
-								public String[] allowableValues() {
-									return schema.allowableValues();
-								}
-
-								@Override
 								public String defaultValue() {
 									return getArrayDefaultValue(parameterName, pageableDefault, schema.defaultValue());
 								}
-
-								@Override
-								public String discriminatorProperty() {
-									return schema.discriminatorProperty();
-								}
-
-								@Override
-								public DiscriminatorMapping[] discriminatorMapping() {
-									return schema.discriminatorMapping();
-								}
-
-								@Override
-								public boolean hidden() {
-									return schema.hidden();
-								}
-
-								@Override
-								public boolean enumAsRef() {
-									return schema.enumAsRef();
-								}
-
-								@Override
-								public Class<?>[] subTypes() {
-									return schema.subTypes();
-								}
-
-								@Override
-								public Extension[] extensions() {
-									return schema.extensions();
-								}
-
-								@Override
-								public AdditionalPropertiesValue additionalProperties() {
-									return schema.additionalProperties();
-								}
 							};
 						}
-
-						@Override
-						public int maxItems() {
-							return arraySchema.maxItems();
-						}
-
-						@Override
-						public int minItems() {
-							return arraySchema.minItems();
-						}
-
-						@Override
-						public boolean uniqueItems() {
-							return arraySchema.uniqueItems();
-						}
-
-						@Override
-						public Extension[] extensions() {
-							return arraySchema.extensions();
-						}
 					};
-				}
-
-				@Override
-				public Content[] content() {
-					return parameter.content();
-				}
-
-				@Override
-				public boolean hidden() {
-					return parameter.hidden();
-				}
-
-				@Override
-				public ExampleObject[] examples() {
-					return parameter.examples();
-				}
-
-				@Override
-				public String example() {
-					return parameter.example();
-				}
-
-				@Override
-				public Extension[] extensions() {
-					return parameter.extensions();
-				}
-
-				@Override
-				public String ref() {
-					return parameter.ref();
 				}
 			};
 			return Optional.of(parameterNew);
